First iteration of Acl methods and web structuring.
This is the first iteration of the basic Acl management idea. A
basic web ui is added to modify Acls too. Acls should theoretically
work as follows:
    A user may be associated with a role. The user inherets the
    permissions of that role. If no extra affiliations are added,
    the permissions only count for keys that the user owns. A user
    can have multiple roles, and with the help of affiliations, a
    scenario where a user with no keys can have access to corp A's
    wallet journal and corp B's asset list only is possible.

    Of course, a super user account also exists. This user inherits
    *all* permissions for all affiliations.

The basic idea for structiring the web UI is also introduced in this
commit. Things like layout grinds, view composers for the sidebar and
validation via Form Requests are used.