feat(queues): Reorganize queues by scope
In order to improve jobs balancing and tracking, this is reorganizing
queues by scope. Auto balancing is becoming the default value and
workers amount is starting at 5 (one worker per queue).

Both default
and notifications queue shouldn't expect an heavy load. As a result, it
should not impact server as it has been the case during 3.x starting
life.