feat(versions): Switch package versions display to generic list
Instead having separate place to be able to review package version, this
will provide all information into the same location, for both core and
plugin packages.

This will depends on a Provider update which have to
be inherited by all SeAT package provider in order to be able to provide
proper information and implement auto-discovery.

This SeAT provider
should be able to return enough information to be able to build some
kind of marketplace and make plugin installation easier.

Use packagist API to check if or not an installed package is outdated.

Add a button near versions tab which is generating and storing into
clipboard a markdown formatted table with all package vendor, name and
versions from installed SeAT packages.

This is providing two ways to fetch and build a changelog for a
specified package :
 - Using directly an uri which is targeting a
Markdown formatted file
 - Using a releases/tags API (ie: github
API)

The resulting changelog is then accessible by a button available
near a package name.