feat(ACL): draft some new role UI including category, label and description
Role can now be customized by providing a description and/or a logo.
They can also be renamed.

Permission targeting character or corporation scope can now use a filter
mechanics in order to limit user access. It's an improvement of the old
"affiliation" system which is also allowing to limit character or corp
access for a complete corporation or alliance entity.