fix: prevent exception when user change main character related to an existing account
instead showing an system exception, attempt to retrieve user account related to the newly selected main character.

if there is one, attempt to remove it. We can remove an account if he has no more keys attached to it. If the target character is still related to an account with keys, display a nice error message to assist end user.

Closes eveseat/seat#708