refactor(moon reporter): improve loading time by using dedicated report header
previous system was relaying on the whole moon list (around 400k entries) from which we were deducing all entries without any content related.

this new system is relaying on a dedicated moon report header to which we attach content. if a moon doesn't have any report, then, we don't track for itself resulting in faster load.

we also give the ability to remove a report.

Closes eveseat/seat#718, eveseat/seat#741