fix(auth): when a user account is deleted, drop definitely attached tokens
when user accounts are manually deleted, we were keeping attached tokens but soft deleting them.
as a result, if we had to recover the token or if the character was transferred - end user was not able to create a new account and get an exception due to user relation null exception.

since user deletion is an admin manual action, we assume token might be safely removed definitely.
however, we keep track of the action using security logs.

Closes eveseat/seat#793