Implemented filtering based on current selection
The current filter option worked lightly different from what users expect to see. Instead of filtering on a range of moons, it filtered only moons that has all the selected materials, instead of showing a compromised list that the user wanted.

For example, say you want to only see R16,R32 and R64 moons. With the previous setup, this was simply not possible since you would end up with a query like

->andWhere(moon is R16)->andWhere(moon is R32)->andWhere(moon is R64)

Altering the format of the query allows for all the same usecases at it previously provided, as well as support other usecases.