Fix Alliance Tracking search
'tracking' is not a searchable field because it is not a SQL column.  This fixes the search on the Alliance Tracking page  so its functional again.

Also to note, this should resolve https://github.com/eveseat/seat/issues/838 since the tracking field is no longer orderable (as of commit b2e6520 in 2023) and now not searchable (although it never worked)